Practice delivered more sensibly
Previously, the practice function would choose a random sentence from the paragraph, and it would recalculate the delimiters each time. Now, the practice button cycles through the sentences it gives in a straightforward manner, and it saves the split sentences in memory so the calculation only needs to be carried out once.

Minor css change as well so that the control panel is centered.